build: split integration and release workflows into gradle scripts
